Move canHandle check for KVMNonManagedStorageStrategy class and fix for failing local storage migrations
