+++ engine/storage/datamotion/src/test/java/org/apache/cloudstack/storage/motion/StorageSystemDataMotionStrategyTest.java
@@ -18,11 +18,9 @@
  */
 package org.apache.cloudstack.storage.motion;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 import static org.mockito.Mockito.doReturn;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 import static org.mockito.MockitoAnnotations.initMocks;
 
 import java.util.HashMap;
@@ -50,17 +48,13 @@ import org.mockito.runners.MockitoJUnitRunner;
 
 import com.cloud.agent.api.MigrateCommand;
 import com.cloud.agent.api.to.VirtualMachineTO;
-import com.cloud.host.Host;
 import com.cloud.host.HostVO;
-import com.cloud.hypervisor.Hypervisor;
 import com.cloud.storage.DataStoreRole;
 import com.cloud.storage.ImageStore;
-import com.cloud.storage.ScopeType;
 import com.cloud.storage.Storage;
 import com.cloud.storage.Storage.StoragePoolType;
 import com.cloud.storage.Volume;
 import com.cloud.storage.VolumeVO;
-import com.cloud.utils.exception.CloudRuntimeException;
 
 @RunWith(MockitoJUnitRunner.class)
 public class StorageSystemDataMotionStrategyTest {
@@ -79,38 +73,6 @@ public class StorageSystemDataMotionStrategyTest {
     private ImageStore destinationStore;
     @Mock
     private PrimaryDataStoreDao primaryDataStoreDao;
-    @Mock
-    VolumeInfo volumeInfo1;
-    @Mock
-    VolumeInfo volumeInfo2;
-    @Mock
-    DataStore dataStore1;
-    @Mock
-    DataStore dataStore2;
-    @Mock
-    DataStore dataStore3;
-    @Mock
-    StoragePoolVO pool1;
-    @Mock
-    StoragePoolVO pool2;
-    @Mock
-    StoragePoolVO pool3;
-    @Mock
-    StoragePoolVO pool4;
-    @Mock
-    Host host1;
-    @Mock
-    Host host2;
-
-    Map<VolumeInfo, DataStore> migrationMap;
-
-    private static final Long POOL_1_ID = 1L;
-    private static final Long POOL_2_ID = 2L;
-    private static final Long POOL_3_ID = 3L;
-    private static final Long POOL_4_ID = 4L;
-    private static final Long HOST_1_ID = 1L;
-    private static final Long HOST_2_ID = 2L;
-    private static final Long CLUSTER_ID = 1L;
 
     @Before
     public void setUp() throws Exception {
@@ -120,38 +82,6 @@ public class StorageSystemDataMotionStrategyTest {
         dataObjectDestination = mock(VolumeObject.class);
 
         initMocks(strategy);
-
-        migrationMap = new HashMap<>();
-        migrationMap.put(volumeInfo1, dataStore2);
-        migrationMap.put(volumeInfo2, dataStore2);
-
-        when(volumeInfo1.getPoolId()).thenReturn(POOL_1_ID);
-        when(primaryDataStoreDao.findById(POOL_1_ID)).thenReturn(pool1);
-        when(pool1.isManaged()).thenReturn(false);
-        when(dataStore2.getId()).thenReturn(POOL_2_ID);
-        when(primaryDataStoreDao.findById(POOL_2_ID)).thenReturn(pool2);
-        when(pool2.isManaged()).thenReturn(true);
-        when(volumeInfo1.getDataStore()).thenReturn(dataStore1);
-
-        when(volumeInfo2.getPoolId()).thenReturn(POOL_1_ID);
-        when(volumeInfo2.getDataStore()).thenReturn(dataStore1);
-
-        when(host1.getId()).thenReturn(HOST_1_ID);
-        when(host1.getClusterId()).thenReturn(CLUSTER_ID);
-        when(host1.getHypervisorType()).thenReturn(Hypervisor.HypervisorType.KVM);
-        when(host2.getId()).thenReturn(HOST_2_ID);
-        when(host2.getClusterId()).thenReturn(CLUSTER_ID);
-        when(host2.getHypervisorType()).thenReturn(Hypervisor.HypervisorType.KVM);
-
-        when(dataStore1.getId()).thenReturn(POOL_1_ID);
-        when(pool1.getPoolType()).thenReturn(Storage.StoragePoolType.NetworkFilesystem);
-        when(pool2.getPoolType()).thenReturn(Storage.StoragePoolType.NetworkFilesystem);
-        when(pool2.getScope()).thenReturn(ScopeType.CLUSTER);
-
-        when(dataStore3.getId()).thenReturn(POOL_3_ID);
-        when(primaryDataStoreDao.findById(POOL_3_ID)).thenReturn(pool3);
-        when(pool3.getPoolType()).thenReturn(Storage.StoragePoolType.NetworkFilesystem);
-        when(pool3.getScope()).thenReturn(ScopeType.CLUSTER);
     }
 
     @Test
@@ -276,73 +206,4 @@ public class StorageSystemDataMotionStrategyTest {
         }
     }
 
-    @Test
-    public void testVerifyLiveMigrationMapForKVM() {
-        ((StorageSystemDataMotionStrategy) strategy).verifyLiveMigrationForKVM(migrationMap, host2);
-    }
-
-    @Test(expected = CloudRuntimeException.class)
-    public void testVerifyLiveMigrationMapForKVMNotExistingSource() {
-        when(primaryDataStoreDao.findById(POOL_1_ID)).thenReturn(null);
-        ((StorageSystemDataMotionStrategy) strategy).verifyLiveMigrationForKVM(migrationMap, host2);
-    }
-
-    @Test(expected = CloudRuntimeException.class)
-    public void testVerifyLiveMigrationMapForKVMNotExistingDest() {
-        when(primaryDataStoreDao.findById(POOL_2_ID)).thenReturn(null);
-        ((StorageSystemDataMotionStrategy) strategy).verifyLiveMigrationForKVM(migrationMap, host2);
-    }
-
-    @Test(expected = CloudRuntimeException.class)
-    public void testVerifyLiveMigrationMapForKVMMixedManagedUnmagedStorage() {
-        when(pool1.isManaged()).thenReturn(true);
-        when(pool2.isManaged()).thenReturn(false);
-        ((StorageSystemDataMotionStrategy) strategy).verifyLiveMigrationForKVM(migrationMap, host2);
-    }
-
-    @Test
-    public void canHandleKVMLiveStorageMigrationSameHost() {
-        StorageSystemDataMotionStrategy st = ((StorageSystemDataMotionStrategy) strategy);
-        StrategyPriority priority = st.canHandleKVMNonManagedLiveStorageMigration(migrationMap, host1, host1);
-        assertEquals(StrategyPriority.CANT_HANDLE, priority);
-    }
-
-    @Test
-    public void canHandleKVMLiveStorageMigrationInterCluster() {
-        StorageSystemDataMotionStrategy st = ((StorageSystemDataMotionStrategy) strategy);
-        when(host2.getClusterId()).thenReturn(5L);
-        StrategyPriority priority = st.canHandleKVMNonManagedLiveStorageMigration(migrationMap, host1, host2);
-        assertEquals(StrategyPriority.CANT_HANDLE, priority);
-    }
-
-    @Test
-    public void canHandleKVMLiveStorageMigration() {
-        StorageSystemDataMotionStrategy st = ((StorageSystemDataMotionStrategy) strategy);
-        StrategyPriority priority = st.canHandleKVMNonManagedLiveStorageMigration(migrationMap, host1, host2);
-        assertEquals(StrategyPriority.HIGHEST, priority);
-    }
-
-    @Test
-    public void canHandleKVMLiveStorageMigrationMultipleSources() {
-        StorageSystemDataMotionStrategy st = ((StorageSystemDataMotionStrategy) strategy);
-        when(volumeInfo1.getDataStore()).thenReturn(dataStore2);
-        StrategyPriority priority = st.canHandleKVMNonManagedLiveStorageMigration(migrationMap, host1, host2);
-        assertEquals(StrategyPriority.HIGHEST, priority);
-    }
-
-    @Test
-    public void canHandleKVMLiveStorageMigrationMultipleDestination() {
-        StorageSystemDataMotionStrategy st = ((StorageSystemDataMotionStrategy) strategy);
-        migrationMap.put(volumeInfo2, dataStore3);
-        StrategyPriority priority = st.canHandleKVMNonManagedLiveStorageMigration(migrationMap, host1, host2);
-        assertEquals(StrategyPriority.HIGHEST, priority);
-    }
-
-    @Test
-    public void testCanHandleLiveMigrationUnmanagedStorage() {
-        when(pool2.isManaged()).thenReturn(false);
-        StorageSystemDataMotionStrategy st = ((StorageSystemDataMotionStrategy) strategy);
-        StrategyPriority priority = st.canHandle(migrationMap, host1, host2);
-        assertEquals(StrategyPriority.HIGHEST, priority);
-    }
 }
